Correcciones

+++ app/controller/product.controller.php
@@ -3,6 +3,7 @@
 // Mac OS
 require_once 'app/model/product.model.php';
 require_once 'app/view/product.view.php';
+require_once 'app/helpers/auth.helper.php';
 
 // Windows OS
 
@@ -10,11 +11,17 @@ class ProductController{
 
     private $model;
     private $view;
+    private $helper;
+
 
     public function __construct(){
-        session_start();
         $this->model = new ProductModel();
         $this->view = new ProductView();
+        $this->helper = new AuthHelper();
+
+        if(session_status() != PHP_SESSION_ACTIVE){
+            session_start();
+        }
     }
 
     public function showHome(){
@@ -27,108 +34,78 @@ class ProductController{
     }
 
     public function showProduct($id_product){
-        $product = $this->model->getProduct($id_product);
-        $this->view->showProduct($product);
-    }
 
-    public function showCategories(){
-        $categories = $this->model->getCategories();
-        $this->view->showCategories($categories);
-    }
-
-    public function showCategory($id_category){
-        $array_products = $this->model->getCategory($id_category);
-        $this->view->showProducts($array_products);
+        if(isset($id_product)){
+            $product = $this->model->getProduct($id_product);
+            $this->view->showProduct($product);
+        }
     }
 
 
     // ABM product
 
     public function formAddProduct(){
-        $categories = $this->model->getCategories();
+
+        $this->helper->checkLoggedIn();
+
+        $categories = $this->model->formAddProduct();
         $this->view->addProduct($categories);
     }
 
     public function addProduct(){
 
-        $nombre = $_POST['nombre'];
-        $precio = $_POST['precio'];
-        $categoria = $_POST['categoria'];
-        $descripcion = $_POST['descripcion'];
+        $this->helper->checkLoggedIn();
 
-        $id_producto = $this->model->addProduct($nombre, $precio, $categoria, $descripcion);
+        if(isset($_POST['nombre']) && isset($_POST['descripcion']) && isset($_POST['precio']) && isset($_POST['categoria'])){
+            $nombre = $_POST['nombre'];
+            $precio = $_POST['precio'];
+            $categoria = $_POST['categoria'];
+            $descripcion = $_POST['descripcion'];
+    
+            $id_producto = $this->model->addProduct($nombre, $precio, $categoria, $descripcion);
+        }
 
         header("Location: " . BASE_URL); 
 
     }
 
     public function deleteProduct($id_product){
-        $this->model->deleteProduct($id_product);
+
+        $this->helper->isLogged();
+
+        if(isset($id_product)){
+            $this->model->deleteProduct($id_product);
+        }
+
         header("Location: " . BASE_URL);
 
     }
 
     public function formModifyProduct($id_product){
-        $this->view->modifyProduct($id_product);
+
+        $this->helper->isLogged();
+
+        if(isset($id_product)){
+            $this->view->modifyProduct($id_product);
+        }
         
     }
 
     public function modifyProduct(){
 
-        $nombre = $_POST['nombre'];
-        $precio = $_POST['precio'];
-        $categoria = $_POST['categoria'];
-        $descripcion = $_POST['descripcion'];
-        $id = $_POST['id'];
+        $this->helper->isLogged();
 
-        $this->model->modifyProduct($id, $nombre, $precio, $categoria, $descripcion);
-
-        header("Location: " . BASE_URL);
-    }
-
-
-    // ABM category
-
-    public function formAddCategory(){
-        $this->view->AddCategory();
-    }
-
-    public function addCategory(){
-
-        $nombre = $_POST['nombre'];
-        $descripcion = $_POST['descripcion'];
-
-        $id_category = $this->model->addCategory($nombre, $descripcion);
-
-        header("Location: " . BASE_URL); 
-
-    }
-
-    public function deleteCategory($id_category){
-        $this->model->deleteCategory($id_category);
-        header("Location: " . BASE_URL);
-
-    }
-
-    public function formModifyCategory($id_category){
-        $this->view->modifyCategory($id_category);
-        
-    }
-
-    public function modifyCategory(){
-
-        $nombre = $_POST['nombre'];
-        $descripcion = $_POST['descripcion'];
-        $id = $_POST['id'];
-
-        $this->model->modifyCategory($id, $nombre, $descripcion);
+        if(isset($_POST['nombre']) && isset($_POST['descripcion']) && isset($_POST['precio']) && isset($_POST['categoria'])){
+            $nombre = $_POST['nombre'];
+            $precio = $_POST['precio'];
+            $categoria = $_POST['categoria'];
+            $descripcion = $_POST['descripcion'];
+            $id = $_POST['id'];
+    
+            $this->model->modifyProduct($id, $nombre, $precio, $categoria, $descripcion);
+        }
 
         header("Location: " . BASE_URL);
     }
-
-
-
-
-
 
 }
